feat: add update action

+++ app/src/main/java/com/example/hw_room/ui/users/UsersFragment.kt
@@ -1,5 +1,6 @@
 package com.example.hw_room.ui.users
 
+import android.app.AlertDialog
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -7,11 +8,15 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.recyclerview.widget.LinearLayoutManager
+import com.example.hw_room.R
 import com.example.hw_room.adapter.UserAdapter
 import com.example.hw_room.databinding.FragmentUsersBinding
 import com.example.hw_room.listener.ItemClickListener
 import com.example.hw_room.model.User
 import com.example.hw_room.repository
+import com.example.hw_room.utils.RESULT_KEY
+import com.example.hw_room.utils.USER_RESULT_KEY
+import com.example.hw_room.utils.addBottomSpaceDecorationRes
 
 class UsersFragment : Fragment(), ItemClickListener {
 
@@ -38,6 +43,14 @@ class UsersFragment : Fragment(), ItemClickListener {
         super.onViewCreated(view, savedInstanceState)
         initRecycler()
         observeUsers()
+
+        childFragmentManager.setFragmentResultListener(RESULT_KEY, this) { _, bundle ->
+            val newUser =
+                bundle.getSerializable(USER_RESULT_KEY) as? User ?: return@setFragmentResultListener
+
+            viewModel.updateUser(newUser)
+        }
+
     }
 
     override fun onDestroyView() {
@@ -55,10 +68,32 @@ class UsersFragment : Fragment(), ItemClickListener {
         recycler.apply {
             layoutManager = LinearLayoutManager(requireContext())
             adapter = userAdapter
+            addBottomSpaceDecorationRes(resources.getDimensionPixelSize(R.dimen.recycler_bottom_padding))
         }
     }
 
-    override fun onDeleteItemClickListener(user: User) {
-        viewModel.deleteUser(user)
+    private fun showAlertDialog(user: User) {
+        AlertDialog.Builder(requireContext())
+            .setMessage("Do you really want to delete this record?")
+            .setPositiveButton(R.string.yes) { _, _ ->
+                viewModel.deleteUser(user)
+            }
+            .setNegativeButton(R.string.no, null)
+            .show()
+    }
+
+    private fun showUpdateDialog(user: User) {
+        UpdateDialogFragment.getInstance(
+            user
+        )
+            .show(childFragmentManager, null)
+    }
+
+    override fun onDeleteButtonItemClickListener(user: User) {
+        showAlertDialog(user)
+    }
+
+    override fun onUpdateItemClickListener(user: User) {
+        showUpdateDialog(user)
     }
 }
